// crates/madari-tv/web/src/client/types.gen.ts
export type PlaybackPreferences = {
    audio_commentary?: TrackPreference;
    audio_description?: TrackPreference;
    audio_languages?: Array<string>;
    subtitle_forced?: TrackPreference;
    subtitle_languages?: Array<string>;
    subtitle_sdh?: TrackPreference;
    subtitles_enabled?: boolean;
};
// ...
export type TrackPreference = 'any' | 'prefer' | 'avoid';